I may suck at coding, but this is working to an agree-able level.   Now I just really need to set up a better way of implementing spells to the AI

+++ 404GameUnnamed/Player.h
@@ -46,7 +46,7 @@ private:
 
 	Collider* colObj;
 	DrawableObject* d_object;
-	GLuint health = 100;
+	float health = 100;
 
 	glm::vec3 position;
 	glm::vec3 eye;
